Backend/requirements.txt
google-generativeai
SpeechRecognition
pyttsx3
langdetect
googletrans==4.0.0-rc1
python-dotenv
